Move EmptyPoint to the Point controller; Remove cards mock file
